hele uz jsem nic moc nevykoumal.. behem zitrka (pro TEbe dneska) se mi na to mrkne ta kamaradka tak uvidim.. kdyby jsi na neco prisel dej vedet diky

+++ src/model/Utvar.java
@@ -15,9 +15,9 @@ public interface Utvar {
 
     public void kresli(Graphics g);
 
-    public void nastavZacatek(int x, int y);
+    public void nastavZacatek(double x, double y);
 
-    public void nastavKonec(int x, int y);
+    public void nastavKonec(double x, double y);
     
     
 }
